fix bugs

+++ nmatheg/configs.py
@@ -38,7 +38,7 @@ def create_configs(config, data_config):
     train_config = {'epochs':int(config['train']['epochs']),
                     'save_dir':config['train']['save_dir'],
                     'batch_size':int(config['train']['batch_size']),
-                    'lr':int(config['train']['lr'])}
+                    'lr':float(config['train']['lr'])}
 
     model_config = {'model_name':config['model']['model_name'],
                     'vocab_size':int(config['tokenization']['vocab_size']),
